LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/view-transition-types-mutable-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: Old view transition aborted by new view transition.
CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: Old view transition aborted by new view transition.

Harness Error (FAIL), message = Unhandled rejection: Old view transition aborted by new view transition.

PASS ViewTransition.types is a ViewTransitionTypeSet
PASS ViewTransitionTypeSet behaves like an ordinary Set of strings
PASS ViewTransitionTypeSet should reflect its members for a non-active transition
